pyproject.toml
[project]
name = "quotescape"
version = "1.0.0"
description = "Generate beautiful quote wallpapers for your desktop"
authors = [
    {name = "Andrei Biswas", email = "andrei.biswas@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = "==3.11.*"
keywords = ["wallpaper", "quotes", "kindle", "desktop", "customization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Desktop Environment",
    "Topic :: Utilities",
]

dependencies = [
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "selenium>=4.0.0",
    "beautifulsoup4>=4.11.0",
]

[project.urls]
Homepage = "https://github.com/codeabiswas/quotescape"
Documentation = "https://github.com/codeabiswas/quotescape#readme"
Repository = "https://github.com/codeabiswas/quotescape"
Issues = "https://github.com/codeabiswas/quotescape/issues"

[project.scripts]
quotescape = "quotescape.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/quotescape/**/*.py",
    "assets/fonts/*.ttf",
    "config/quotescape.yaml",
]

[tool.hatch.build.targets.wheel]
packages = ["src/quotescape"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests",
    "types-PyYAML",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=quotescape --cov-report=html --cov-report=term"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__ files

[tool.isort]
profile = "black"
line_length = 100